requirements.txt
pynput
mysql-connector-python